% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe_data.R
\name{enframe_data}
\alias{enframe_data}
\title{Enframe biodiversity data.}
\usage{
enframe_data(x)
}
\arguments{
\item{x}{a list of assemblage data frames}
}
\value{
a data frame with one row per assemblage
}
\description{
\code{enframe_data} returns a data frame with one row per assemblage and a
list column of frequency vectors
}
\examples{
l <- list(
  A = tibble::tibble(A1 = c(10,2,1,1)),
  B = tibble::tibble(B1 = c(7,2,1))
)
enframe_data(l)
# each assemblage data frame must be named
l <- list(
  A = tibble::tibble(A1 = c(10,2,1,1)),
  tibble::tibble(B1 = c(7,2,1))
)
enframe_data(l)
}
