% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_data.R
\name{nest_data}
\alias{nest_data}
\title{nest_data}
\usage{
nest_data(x)
}
\arguments{
\item{x}{a list of assemblage frequency vectors}
}
\value{
a nested data frame with one row per assemblage
}
\description{
nest_data returns a nested data frame with one row per assemblage
}
\examples{
# abundance data
l <- list(A = c(10,2,1,1), B=c(7,2,1))
nest_data(l)
}
